requirements.txt
GitPython==3.1.32
pytest==7.4.0
